Rock Paper Scissor start

+++ client/src/components/settingSectoins/Drawer.js
@@ -25,12 +25,55 @@ export default function PLayersDrawer({allPlayers, currentPlayer, socket}) {
     if (event.target){
       socket.emit('check_player', targetPlayer, (exist) => {
         if(exist){
-          socket.emit('get_roles',(roles) => {
-            const buttons = roles.map((role) => {
-              return {
-                label: `${role} role`,
+          confirmAlert({
+            title: 'Game request',
+            message: `Choose the Game you want to play`,
+            buttons: [
+              {
+                label: 'TicTacToe Game',
                 onClick: () => {
-                  socket.emit('game_request', currentPlayer.username, targetPlayer, role, (tarPLayer) => {
+                  socket.emit('get_roles',(roles) => {
+                    const buttons = roles.map((role) => {
+                      return {
+                        label: `${role} role`,
+                        onClick: () => {
+                          socket.emit('game_request', currentPlayer.username, targetPlayer, role, "TicTacToe",  (tarPLayer) => {
+                            localStorage.setItem('hanging_request', targetPlayer)
+                            confirmAlert({
+                              title: 'Confirm game request',
+                              message: `Waiting ${targetPlayer} response`,
+                              buttons: [
+                                {
+                                  label: 'Cancel',
+                                  onClick: () => {
+                                    localStorage.removeItem('hanging_request');
+                                    socket.emit('cancel_request', targetPlayer);
+                                  }
+                                }
+                              ],
+                              onClickOutside: () => {
+                                localStorage.removeItem('hanging_request');
+                                socket.emit('cancel_request', targetPlayer);
+                              },
+                            });
+                          
+                          }); 
+                        }
+                      }
+                    });
+                    confirmAlert({
+                      title: 'Game request',
+                      message: `Choose the role for the game (role number) is the number of winning required to get next level`,
+                      buttons: buttons
+                    });
+                  });
+                  
+                }
+              },
+              {
+                label: 'Rock Paper Sessior Game',
+                onClick: () => {
+                  socket.emit('game_request', currentPlayer.username, targetPlayer, null, "RPS", (tarPLayer) => {
                     localStorage.setItem('hanging_request', targetPlayer)
                     confirmAlert({
                       title: 'Confirm game request',
@@ -51,14 +94,10 @@ export default function PLayersDrawer({allPlayers, currentPlayer, socket}) {
                     });
                   
                   }); 
+                  
                 }
               }
-            });
-            confirmAlert({
-              title: 'Game request',
-              message: `Choose the role for the game (role number) is the number of winning required to get next level`,
-              buttons: buttons
-            });
+            ]
           });
         };
